datasets/taxi-logs/requirements.txt
pymongo==4.3.3
python-dotenv==0.21.0
tqdm==4.64.1
